Make the PlusGateway able to use Repository
WooCommerce make the Gateway aware of options instead of separating the concerns.
This will prevent us to have a Repository class separated by the gateway to reuse in different
context.

So, since `PayPalPlusGateway` need to get a `PlusRepository` instance and since `PlusRepository`
need to get a `PayPalGateway` instance from which retrieve the option we'll end up into an dependency
loop issue.

The idea to solve this was to move the implementation into a Trait and the constants
into the interface.

This will allow us to reuse the repository internally of PlusGateway.

The downside here is that the Gateway is now implementing a `PlusStorable`
interface but that's the way I found.

PPP-283