Introduce gulp tasks in place of grunt
1. Remove frontend js files that take only space and do nothing
2. PayPalPlus SDK is now within the repo.

This to avoid problems when a package is created because it's possible
a newer version not tested goes into the package, making difficult support
and incorrect package.

Also, since we are using `mozart` and it's a dev dependency having the sdk
as normal dependency will create problems in case the plugin is used as
composer package because the sdk will not be moved under `lib` nor will
get the prefix in case the `--no-dev` is passed to composer.

3. Make the admin.js compatible with js before es6