Introduce compile automation for css and js files
To follow the same hierarchy of MLP i removed `assets` directory and
replaced it with `public`.

The css is now converted to sass, this because even though currently
we have only one rule, by introducing ECS we'll need to add more css
to stylize buttons and will be also useful in future to prevent refactors
for new implementations.

Javascript has been ported to es6/es2015 because it's more simple to
bundle everything by using `modules` and have more advantages because of
the new features we can use.

Webpack is used to bundle the things.
I haven't just used babel because it doesn't include the module require
logic, where webpack do and it's what we need.

Css are not converted by webpack because of problems and unnecessary
complexity in maintain the tool. Webpack get the scss convert it into
css and produce a css that is javascript, is possible to create a separated
bundle for css but I don't thing it's really needed.

Since we probably want to have the compiled css into the repository so
everyone that will include the plugin as a composer package can just use it
without the needs to compile assets.

This is another reason why `admin.js` is under repository.

PPP-301