Do not differentiate between assets and minified versions
Browsers works pretty well with source maps, so for development purpose
we don't need to have original file and compressed one, the compressed
it's ok if we have a source map, just have `resources` directory into your
env server and everything will works fine.

On production we don't need to ship the source map, so I removed the
`SCRIPT_DEBUG` conditional to load different files.

In case it's needed by the support we could include the source map later

PPP-301