Doing the payment return the newly object instead of a boolean
Returning a boolean isn't useful at all because doesn't give us any
information about the status of the payment.

When the Api1 do a execute payment the response from the api will replace
the information into the actual object, so returning it will allow us
to inspect the result.

PPP-306