Fix wrong price sent to PayPal API because of tax calculation
WooCommerce allow shop managers to define the prices including or excluding
taxes.

Internally the platform will keep the price and the tax separated and will
do the correct math afterwards.

Our plugin instead of retrieving the product price will get the subtotal
that doesn't include the tax this created the problem discussed in PPP-40.

What we want instead it's always send the price WooCommerce calculate with
or without taxes based on his settings.

Practically we'll send always the price including taxes because PayPal
want to do the math and the total should be the sum of the product prices
+ taxes and/or shipping taxes +/- discount + fees.

We have to tell to PayPal which is the price of the product indipendently
if this price include or not taxes, taxes are calculated separately and will
be included based on `wc_prices_include_tax` returned value.
See `\WCPayPalPlus\Payment\PatchProvider::get_payment_amount_patch` for
more information.

Also, fixed an issue for rounding values because of a wrong decimal
calculation depended on currency decimal support.
See `PriceFormatterTrait`.

PPP-40