Delete redundant logs
The logger we are using is already hooked into PayPal Api Configuration
and the logs are stored into WooCommerce files.

Having to log `PayPalConnectionException` is redundant.

Also, fixed a type issue for ErrorData details, expected a json string
but got an array when extract the data by the Json.

PPP-336