Address Rounding problems
The main problem here is how WooCommerce tackle the taxes applied to prices and how
PayPal needs to have those prices.

Because of intrinsic rounds problems WooCommerce deal with taxes as cent values, means
the numbers used to calculate taxes contains a lot of information.

PayPal want to have prices 2 decimal rounded, if not the Item::setPrice will format in that way.
What we did in previous versions was to get the price of the product (total) divided by
the quantity see $data->get_price(), this could produce rounding problems because
the division result could contains more than 2 decimal points, then we pass that value
to Item::setPrice that will round it, practically adding some more cents to the price.

So the price calculated by WooCommerce will not fit the price calculated by our implementation.
The good solution would work as WooCommerce (using cent values) but actually the logic
is too complicated that need a complete separated implementation.

So as workaround to send to paypal all the items instead only one as done in previous
versions we implemented the following solution.

PPP-40