Payment Data task isn't executed because of a problem with WooCommerce and Nonce Validation
In case of a non logged in user the nonce validation could fail because
WooCommerce has a filter to `nonce_user_logged_out` that interferer to
nonce validation after the cart is update because the WooCommerce session
isn't create before.

Solution is to remove the filter while we are validating the nonce
and add it again after.

This will ensure WooCommerce will not interferer during our nonce validation.

PPP-345