Add order Note on payment status change to refund
WooCommerce change the status to refund by it self before we are able
to update the order during the IPN request.

That's means there's not status transition and the note
"Payment Refunded via IPN." will not be added in the order.

So since WooCommerce take care of the status transition what we have
to do it's just add an order note.

PPP-364