Render Express Checkout Payments Button on `updated_shipping_method` event
There are more places where the express checkout button is printed and
depending on the place WooCommerce could make ajax calls which change
the markup structure.

When that happen the payment button is removed from the DOM this because
there's no more a reference to an element in the dom.

When then the markup is recreated PPP does not trigger the render of the
button.

There is a list of events in the `expressCheckout` that I ported into
php env to allow third parties to filter that list in this way possible
renders needed because of other WooCommerce parts or other plugins can be
addressed at support level, than we can add those in a future release.

Basically will not block the merchant to have a functional button in case
of cases we haven't considered yet.

PPP-386