Move service cache inside factory
PPP-409
Instead of throwing to log later we log here
and return null if something went wrong.
The option to check if caching is enabled
is used inside the factory as outside might
have not been defined when we need it.
On fresh installs the shared options might
not be set yet when we call so a default
value is added and caching is not allowed.