Add null object pattern
PPP-409
RemoteResourcesStorer implements the new
RemoteResourcesStorerInterface and the new
NullRemoteResourcesStorer too, this is the null object we pass in case
the option is not enabled or there was an
 error with the filesystem.A factory
is used to decide which one to create based
on the script caching option and the
positive creation of the filesystem. This
way clients will receive the expected object
with or without the expected behavior.
So there is rollback in all the previous
modifications to check the option, as there
is no need. Classes using
RemoteResourcesStorer now use the interface